+++ src/services/ChatService.php
@@ -12,8 +12,7 @@
 namespace doublesecretagency\sidekick\services;
 
 use Craft;
-use craft\errors\MissingComponentException;
-use doublesecretagency\sidekick\constants\Session;
+use craft\helpers\Json;
 use doublesecretagency\sidekick\models\ChatMessage;
 use doublesecretagency\sidekick\models\FunctionCall;
 use yii\base\Component;
@@ -21,64 +20,59 @@ use yii\base\Component;
 class ChatService extends Component
 {
     /**
-     * Clear the existing conversation from the session.
+     * Get the cache key based on the current session.
+     */
+    public function getCacheKey(): string
+    {
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $sessionId = Craft::$app->getSession()->getId();
+        return "chathistory-{$sessionId}";
+    }
+
+    /**
+     * Clear the existing conversation from cache.
      */
     public function clearConversation(): void
     {
-        try {
+        // Delete the chat history
+        /** @noinspection NullPointerExceptionInspection */
+        $deleted = Craft::$app->getCache()->delete($this->getCacheKey());
 
-            // Get the session service
-            $session = Craft::$app->getSession();
-
-            // Remove the Responses API ID
-            $session->remove(Session::RESPONSE_ID);
-
-            // Clear the conversation from the session
-            $session->remove(Session::CHAT_HISTORY);
-
-        } catch (MissingComponentException $e) {
-
-            // Log an error message
-            Craft::error("Unable to clear conversation from the session.", __METHOD__);
-
+        // If not deleted, log an error
+        if (!$deleted) {
+            Craft::error("Unable to clear the chat history.", __METHOD__);
         }
     }
 
     /**
-     * Retrieve the existing conversation from the session.
+     * Retrieve the existing conversation from cache.
      *
      * @return array
      */
     public function getConversation(): array
     {
-        try {
+        // Get the chat history
+        /** @noinspection NullPointerExceptionInspection */
+        $chatHistory = Craft::$app->getCache()->get($this->getCacheKey());
 
-            // Get the existing conversation from the session
-            $conversation = Craft::$app->getSession()->get(Session::CHAT_HISTORY) ?? [];
-
-            // If not a valid conversation
-            if (!$conversation || !is_array($conversation)) {
-                // Return an empty conversation
-                return [];
-            }
-
-            // Return the complete conversation
-            return $conversation;
-
-        } catch (MissingComponentException $e) {
-
-            // Log an error message
-            Craft::error("Unable to get conversation from the session.", __METHOD__);
-
-            // Return an error message
-            return [
-                new ChatMessage([
-                    'role' => ChatMessage::ERROR,
-                    'message' => "Unable to load the conversation."
-                ])
-            ];
-
+        // If no chat history, return an empty array
+        if (!$chatHistory) {
+            return [];
         }
+
+        // Return the JSON decoded chat history
+        return Json::decodeIfJson($chatHistory);
+
+//        // Log an error message
+//        Craft::error("Unable to get the chat history.", __METHOD__);
+//
+//        // Return an error message
+//        return [
+//            new ChatMessage([
+//                'role' => ChatMessage::ERROR,
+//                'message' => "Unable to load the conversation."
+//            ])
+//        ];
     }
 
     /**
@@ -88,21 +82,24 @@ class ChatService extends Component
      */
     public function addMessage(ChatMessage|FunctionCall $message): void
     {
-        // Track the message
-//        Craft::info("Appending message to the conversation history.", __METHOD__);
-
-        // Get the existing conversation from the session
+        // Get the existing conversation from cache
         $conversation = $this->getConversation();
 
         // Append the new message to the conversation
         $conversation[] = $message;
 
-        try {
-            // Save the updated conversation to the session
-            Craft::$app->getSession()->set(Session::CHAT_HISTORY, $conversation);
-        } catch (MissingComponentException $e) {
-            // Log an error message
-            Craft::error("Unable to save updated conversation to the session.", __METHOD__);
-        }
+        // Set lifespan for the cache entry
+        $lifespan = (60 * 60 * 24 * 14); // 14 days
+
+        // Update the chat history
+        /** @noinspection NullPointerExceptionInspection */
+        Craft::$app->getCache()->set(
+            $this->getCacheKey(),
+            Json::encode($conversation),
+            $lifespan
+        );
+
+//        // Log an error message
+//        Craft::error("Unable to update the conversation.", __METHOD__);
     }
 }